Remove the tight coupling between leader Service and IPService/Systemd Service. (#113)
* Remove the tight coupling between leader Service and IPService/Systemd
Service. Instead go for the observer pattern coupling these.

* VIP Service is a service to handle Virtual IPs in the cluster.
Leader  service deals with leadership only. It is decoupled with
observers. Now its possible to have many leaderservices happening on the
system.

for this commit - systemd service, VIP service use the same leader service.
these will be split later when we have separated the APIgw and kubernetes systemd

* Rename LeaderEventData to LeaderEventType per review comments.
Make that a string to make it easier to print while debugging

* Seperate out systemd service, node service, master service
master service depends on both leader service as well as systemd service
Have a env.LeaderService to be used by VIP and Master Service
Have a env.systemdService to be use by Master and Node Service
renamed the systemd/configs to systemd-configs.

* Address the review comments -
Add options to Mock most of the services.

Fix the path to store the cluster object in the etcd.
Now After creating the cluster, a GET using the Api Gateway returns
correct object.

Add static-glibc and golang 1.8.3 to the build-container

* Changes per the review comments.
Do APIServer Config as part of Master Service instead of Node Service.
build-container uses alpine instead of centos : much smaller image

Yet to add tests

* CGO_ENABLED=0 is removed from the Makefile because we might be including
some code with cgo

Added a base-container (from alpine3.6) with kubelet.
This is the new base for the pen-base container instead of centos

removed calling systemctl from initscript since alpine does not have
that executable. Instead call it from CMD on startup with API.

build-container with Go tools also moved to alpine.

utils/systemd exposes an interface and the caller use an interface
instead of directly calling global functions.

Add mock objects for the cmd/services.
This allows us to write unit-tests by mocking out and test the
core-logic.

* Renamed variables as per review comments.
rebase to top.