Notify observers of Leadership change event only when there is a real change. (#148)
* Notify observers of Leadership change event when there is a real change.

When stopping a leaderservice, when the event channel is closed, we
should not do a stop/start again. Instead note that its being stopped
and return from the go routine.

Instead of blindly telling the observers during the 'stop', do it as
part of the ctx cancellation in the go routine.

All these to take care of intermediate failures in #141

* call notification explicitly once.
If the Stop() was called dont send any further notifications.