Makefile,box*.rb: Improve performance/ease of use on docker targets
Right now the copy that's involved is a pretty serious hit now that the
nic, etc repo is there.

As a result, the make target `docker-test` took several minutes to run;
and `docker-test-debug` took even longer to just spawn a shell.

Now, the copy only happens conditionally, which is necessary for CI.
However, it is not for end-user development so the copy is switched off
during manual make tasks.

I also renamed `docker-test-debug` to `test-debug` to keep it consistent
with other environments.

I hope this makes simulating the CI environment more useful for you.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>