KeyMgr first commit (#1054)
The purpose of the diff is to isolate key management functions (creation,
storage and usage) from the components that exercise those functions. For
example, CA generates keys and signs certificates, rpckit client/server do
TLS handshakes.
The PR introduces a new component called KeyMgr to provide this functionality.
KeyMgr has an interface to create/store/use keys, similar to an HSM and a
pluggable backend to implement the operations. Included in the PR is a simple
backend based on Go crypto library for key functions and the utils/certs library
for file-backed storage. This works as a reference and should suffice for
development and testing purposes.
We will soon also have a keymgr backend capable of performing crypto operations
using any device that implements the PKCS#11 standard interface (HSMs,
smartcards, etc.) without exposing the keys to Venice components.

KeyMgr relies on backend for persistence, so it also uses it to store non-
sensitive material like certificates and pulbic keys. It contains a cache of
recently used objects that is populated when miss occurs.