update Leaser status in cluster.Status asynchronously on leadership change (#1616)
* In one case of network partition, when the leader was in the minority of
quorum nodes, the majority was not able to elect a leader.
Add a log to help debug this.

* if cmd is waiting for k8s api server for too long, print a more verbose
    error message

* Problem case:
When network is partitioned and old leader, apiserver are no longer
reachable, the new CMD leader was trying to update the Status object by
contacting the API server. This does not return for 900+ seconds till
the transport is closed. CMD was waiting for this to start the k8s
services (which inturn start the apiserver).

Fix:
update the cluster object with the new leader in status asynchronously in a
go-routine. (other than user) no other code should use this status field to
determine the leader. Instead it should be determined by kvstore
election. so, it should be ok to be updated asynchrously.

Also set a reasonable timeout for the apiclient operations

* updateLeaderinCLusterStatus() will loop till its successful (or)
leadership is gone.

in resolver - check for conditions before adding a service instance as healthy

* Reduce failure detection times during network partitioning.

based on the review comments, have a goroutine which in a poll keeps
updating the Cluster.Status object if it has wrong status

Add time for last leadership update time in the cluster status object.

Handle the pen-cmd restart when the node names are ip address (and the
hostname does not dns resolve to the ip address provided).

When the cluster has network partition and some nodes become
unreachable, detect this loss sooner instead of default k8s timeouts
which are order of minutes. Now the detection is within 20 seconds

* cleanup for loop for readability

* dont try to resolve dns in cmd for hostnames.
pass it as-is to etcd.