Add dol testcase to test congestion avoidance algorithm
Add a testcase to test the logic in tcp_cong_avoid_ai (cwnd is
incremented by 1 for every cwnd worth of data acked). The testcase
starts with snd_cwnd=10 and snd_cwnd_cnt=9. As part of the testcase one
packet worth of data is acked causing snd_cwnd_cnt to increment by 1,
and the following logic to execute:

        tp->snd_cwnd_cnt += acked
        if (tp->snd_cwnd_cnt >= w) {
                u32 delta = tp->snd_cwnd_cnt / w;

                tp->snd_cwnd_cnt -= delta * w;
                tp->snd_cwnd += delta;
        }

Since acked is always 1 for BIC, it appears this part of the logic can
be written without div or mul operators. However using that to test out
the div and mul instructions for now.