update packer-template to eec88a56a04bcaa5f7e0f3e2e089670716c3d707 of github.com/kaorimatz/packer-templates . (#1925)
update to centos packer scripts. (Looks like we were already using centos-7.4 binaries in the name of centos-7.3)
Move to latest epel repo.
Update go to 1.9.2.

centos-7.4 does not work with older virtualboxes. So add a check for 5.1.30 virtualbox version in Vagrantfile

Fix the hang bug in nic_admission_test - Thats happening because the vboxsf file system does not support unixdomain sockets and creating an integration seems to be using it