tsdb transmitter to write to influx (#1873)
* tsdb support to log to real influx from network agent.
utils/debug has come convenience wrappers for logging to tsdb.
    When creating a debug stats, can specify whether to log to tdsb..

fix in agent under race condition when cmdif is not ready before its
being referenced.

fix for crash when watcher is closed (and another thread is accessing
it).

misc fixes

* tests for Batch collector to complete code coverage in tsdb.

minor changes in nmd to make govet happy

* More code coverage

* handle the closing of agent watcher streams better in NPM

* fix endpoint and security watch handlers

* When connection to agents is gone, close the channel and return from the
goroutine.

* include pprof to give good useful http debug bindings  for performance

* Dont strip the executables for now.
If we strip executables, we dont get profiling data to debug in the cpu
usage in the scale tests,

* register clear handler for expvar at /debug/vars/clear

*  add /debug/pprof REST endpoint handler to get profiling data from cmd

* increase the channel capacity to absorb a little bursts and close the
channel when watchers return.

* change from mutex to RWMutex lock so that multiple resolver clients can be
serviced in parallel.

* Locking fix for npm crash.
When a grpc.Dial fails and creation of RPCClient fails, the apps retry
creating the RPCClient. They dont necessarily specify a new balancer and
can reuse one balancer for the same service.
In such a case, the Close() has to cleanup properly the first instance
so that no lingering goroutines can meddle with the second life of
RPCClient.

* locking fixes.
Without these, the agent panics due to concurrent map update.

* handle network/ep/sg events asynchronously.
one of big factors  of scale performance improvements.

* dont fail when client.initRPC fails. network may not be available at
that time (or) the service may have transient failure.

Instead keep retrying periodically.

* Dont fail if we are unable to establish connection to collector at the
start of the day - this can happen due t network failures and/or
collector crashes.
Instead keep retrying periodically to establish the connection

* On Agent side no need to have 4 resolvers to CMD. Use 1 resolver and
share.
Yet another optimization for scale.

* by default set the grpc max message size to 50MB.
this is so that many APIs like List() with say 50k objects get into
payload of the order of 25MB.

* set max size on etcd such that of a BIG get requests, we are not limited
by the size of default data (4MB)

* scale changes:

In memdb.objdb have a seperate lock for watchers and objects.

Minimize code being executed under lock. Aggressively unlock as soon as
posible.

When writing to watchers, use non-blocking write.
Without this one slow (or dead) client will cause backlog to increase (and eventually
the controller is stuck in blocking write to channel affecting traffic
to all the clients.

* scale fixes:
keep one instance of resolver per process. No need to spawn so many
resolver instances through out the code which open a channel to cmd to
listen to updates.

Close client instances when they are done - fixes connection leaks

* scale fixes:

Network watcher returns a list instead of individual event.
At NPM - Just before sending events to agent, check if there are more
events to be sent to agent. If so, collect them and batch up to 100
events together in a message and send them.
If more events are available, send them immediately in a seperate
message. After the last send (of less than 100 messages), wait for 50
msec before trying to send the message again. This way, if more
messsages are on the way, we get a chance to batch them together and
this scheme drastically reduces the numbers of messages sent to agent in
scale conditions.

In the agent, have a queue of messages coming from npm. This way agent
can absorb a burst of messages from npm and work in a separate
goroutine.

* do 'make -C venice/ctrler/npm/rpcserver/netproto' per review comments.
only resulted in cosmetic changes.