CMD certificates server for controllers (#2218)
* CMD certificates server for controllers

This change contains the proto definitions and handlers for a new
certificates API exposed by CMD. This API is meant to be used by
infrastructure components (controllers, agents, ...) to get a
certificate that is trusted within the cluster and can be used
to do TLS with other components.
At the moment the API is not authenticated. Later it will be
restricted such that only local components on a node can request
certificates (after attestation performed by CMD) or remote
components running on NAPLES that have been authenticated and
admitted to the cluster.
The client of this API is rpckit. It automatically gets a
certificate when a component instantiates a new client or server.
The certificates are signed by the CA embedded with CMD using
the signing key that was provided when the CMD node joined
the cluster.

* Autogenerated files for PR #2218