Introduce "test mode" and "default TLS provider" for rpckit (#2211)
Currently if any rpckit client wants to use TLS, it has to explicitly
pass in a provider. With this change, rpckit now has a "default"
TLS provider that is used unless the client passes in a different one.

Right now the default TLS provider is still "nil", i.e. TLS is disabled.
Later it will be switched to "ckmbasedprovider", which is a provider
that is able to generate a CSR for the client, get it signed by a
certificate service running on CMD and return it to rpckit.

TLS should not be enabled for unit-tests. However, we don't want
to have each component or to figure out if it is running in test
mode or not or have each test go and disable it explicitly. At the
same time, we don't want to rely on mechanisms like environment
variables or build tags that prevent developers from simply running
"go test". To provide a single, global knob, we have introduced
a small package called "testenv" which contains flags to control
behavior various components when they are running in test mode.
The default for these flags is "production mode".
Package "testutils", which is included by all tests but not by any
component, switches it to  "test mode" in the init() function.

Note that for integration tests we do want to turn TLS on, so rpckit
has a way to override default TLS provider when running in test mode.

Finally, AES key wrap functions are moved out of keymgr because they
are truely independent components and there's no reason to bundle them
with keymgr.