Add some Venice integration test dependencies to pensando/nic (#2338)
The pensando/nic CI run includes executing Venice integration tests.
Some Venice components have dependencies that are not included in
the pensando/nic box (libtool, softhsm).
Right now integration tests do not cover any component with the
additional dependencies, so everything works fine. However, we
are increasing coverage of the integration tests, so we need the
extra dependencies in the pensando/nic box.