DOL testcase to test timer full
When timer hash table is full, the target doorbell is immediately rung.
So the target program needs to compare the current timestamp with when
the timer was started to actually determine if this is a valid timeout.
However we don't have a concept of current time in the model, so the
best we can check if the timer is immediately rung (in the model we
count on the timer expiry being stepped without having to manually step
the timer). Once rung, we make sure we don't start the timer again to
prevent an endless loop.

To simulate timer full, the hash depth of the timer HBM space is
temporarily reduceed to 2 key lines and the key area is filled with
0xff. Once the testcase finishes, the memory is zeroed out again, and
hash depth is reset.