linux ionic: split driver into two drivers: PCI and platform
For network IO from ARM embedded env, we can use Linux ionic driver
modified as a platform driver and use Device Tree to layout the
device resources.  So split the bottom half of the driver into two
parts: one for tradiational host-based PCI access and one for platform.

This commit adds the code for both drivers and the Makefile can target
building either driver.  The PCI version of the driver is still ionic.ko.
The platform version of the driver is ionic_mnic.ko.