linux ionic: add features cmd
Add new opcode for features cmd.  Feature cmd is use to get features
supported by the device so driver can enable or disable features
in the driver accordingly.

A single feature set is defined for eth hw offload features.  Other
feature sets can be added later, as needed.  The set of eth hw offload
features is based on Linux netdev hw offload features, but is
applicable for any OS wanting eth hw offload features.

The device-side implementation of the features cmd is TDB.  The cmd
is designed to queue on an adminq.  Currently, the Linux driver will
not call the features cmd, so all eth hw offloads are disabled in
the driver.  As features (such as Tx csum offload) are validated in
the device, we can enable the device-side implementation.