Split NIC admission and updates APIs (#2761)
The NIC admission API needs to run on an unauthenticated port because when the
NIC is trying to join the cluster, it does not have a certificate yet.
NIC configuration updates and watches, on the contrary, can run over
authenticated connections because they are only available to NICs that have
already joined the cluster.
This PR splits the existing SmartNIC API into two parts: registration and
updates/watches. The updates/watches part is moved from the CMD unauthenticated
to the authenticated port. It also includes the NMD-side changes needed to
instantiate separate clients and point them to the right endpoint.