linux ionic: improve calling adminq cmds
This patch aims to make queuing adminq cmds more robust by:

1) For context where we can sleep (process context), call adminq cmds
with a wait event to put the calling process to sleep while waiting for
the asynchronous completion.  The completion happens in NAPI context
(softirq context), where the completion wakes the wait event, in turn
waking the original calling process.  The wake event is interruptible
and bound by timeout, in case the user wants to cancel (^C) a command or
the cmd takes too long to complete.  The timeout is set to 2 secs, but
can be adjusted up or down as needed.

2) Use a spin lock to protect queuing cmds on the adminq, since we could
have independent processes trying to queue adminq cmds at the same time.
Basically protect multiple producers queuing to the adminq.

3) In the absence of a real adminq, we call dev cmds.  Move the calling of
the dev cmd to the adminq functions in the shared code, rather than
using #ifdef ADMINQ in the mainline code.  When we get a real adminq, we can
remove the fake dev cmd calls in one place, for all drivers, in the
shared code.  The mainline code should call ionic_adminq_xxx calls and
should be void of any #ifdef ADMINQ wrappers.