Reserve space for hole
Otherwise, client can get unexpected errors posting work requests to the send
or recv queue if the client requests a power-of-two size queue.

Eg: `ibv_rc_pingpong -r4` would fail when posting the fourth recv.

For the completion queue, the the completion queue could be overrun if space is
not reseved for the hole.