linux ionic: move to ionic_api for adminq posting
As part of the RDMA effort, Allen introduced an API into the Linux eth
ionic driver to queue cmds on the adminq, in a way that protects competeing
processes from stepping on each other while queuing on the single adminq,
and also allowing for the calling process to wait (sleep) for the
adminq completion.  (This new API also handles using dev cmds rather than
the real adminq for now, until we get a proper adminq).

The new API was designed for use by the RDMA driver, but it's also useful
for the eth ionic driver to queue cmds on the adminq.  So this patch is
all about moving the queing of adminq cmds to use the new API.  As a result,
the driver code is cleaner and the shared code files ionic_adminq.[ch] are
no longer needed...in fact, they're gone.

For other OSes, the pattern we want to follow is:

1) Implement the equivalent function ionic_api_adminq_post() to post a cmd
on the adminq.  This function needs to hold a lock when putting new cmds
on the adminq.  It also hooks the cmd for a callback, and the callback will
wake the sleeping process.

2) In the driver mainline code, prepare the adminq cmd (on the stack or on
the heap), filling in the cmd fields, and then call ionic_api_adminq_post().
The driver can wait for the cmd completion by putting the process to sleep.
If the process can't sleep (in interrupt context), then the process can
schedule a deferred process to post the cmd and wait for completion.
Either way, once the process is woken, the comletion (status, etc) can
be processed.