Modify existing max_data_rate() to use 8K buffer CP/DC dol
Existing dol that meant to measure the latency by loading the
compression and decompression engine with 16 and 2 requests
respectively using 4K buffers had very minimal checks and needed
additional changes to alter the size of the buffer.

This commit is changing the compress/decompress buffer size alone to
8K to quickly get the performance numbers (although it might be
useful to  manipulate with more sizes, checks, etc.).

Signed-off-by: Mohan Kumar <mohan@pensando.io>