linux ionic: change feature cmd to be two-way street
Modify features adminq cmd.  Now, cmd takes a set of features the driver
wants to enable and the completion returns a set of features the device
supports.  After the completion, the driver should take the AND of the
device supported features with the driver's requested features, and use
the result as the running feature set.

This allows the device behaviour to change based on features wanted by
the driver.  For example, if driver wants VLAN strip enabled, the device
can strip the inline VLAN from the pkt.  If the driver doesn't want VLAN
strip enable, the device would not strip the VLAN tag from the pkt.