Give 10secs delay for RDMA writes(only RTL) to appear in ASIC (#3688)
right after Write in TestSetup() or Write()in TestTeardown() to appear
before the immdt test starts

One problem we have seen due to this is at the end of TestTeardown()
we are resetting the proxy_pindex and proxy_cindex to pindex, and then
issue write_mem_pcie() which is async call and might take long time
before the write appears in HBM. When DOL next test starts immediately,
it will read HBM syncronusly, which still has older proxy_cindex, which
is noted as Q Pre State. At the end of test, it will expect one increment
but test shows incremented by 2 from the Q pre-state, though test only
incremenetd one, and and additional incr is due to delay write from the
prev test.

Sleep of 10 fixed all the open issues in RTL. And this delay is not needed
in DOL with Standalone model, so restricited this sleep to RTL only.