Add tcp_normalize_mss features which will be part of tcp_stateless (#3834)
normalization. If the tcp_normalize_mss value is non-zero as part of the
security profile then all packets with SYN flag set will be inspected
for MSS option and if present the MSS Value will be modified to
tcp_normalize_mss, if the TCP MSS option is not present then the option
will be inserted and all appropirate lengths and tcp_dataoffset will be
updated. Added 3 different DOL Cases to cover these diffeerent cases.
Fixed the bug in tcp_data_len computation when IPv6 options were
present. Had to do this in l4_profile as the input_mapping_native and
input_mapping_tunnel key makers were full.