Support to receive packets with PSN/MSN greater than retried PSN/MSN (#3909)
- Requester should maintain the maximum PSN/MSN advances so that
in case of ack drops, responses to retried request having PSN/MSN
logically greater than PSN/MSN of retried request can be allowed. This
would happen if responder received and ack'ed all requests but acks were
dropped. To handle this case, maintain max_tx_psn/max_ssn in sqcb1 and
update them with the maximum advances of tx_psn and ssn. In the event
of backtracking, these fields are not backtracked and will still
maintain the maximum value so that responses with greater PSN/MSN can
be allowed.