linux ionic: flush rx queue after queue disable
Cleanup any Rx pkts that arrive after interrupts are disabled on the
completion queue but before the queue has not been disabled.  We want
the driver's accounting of the queue ring positions (head, tail) to
match the device's accounting, so that when queue is re-enabled,
everything continues to work.  The race we're closing is this:

     mask intr
     sync intr             <-- Rx pkts arrive on queue
     napi disable
     disable q

After disable queue, we need to flush the Rx pkts and return intr
credits.