Fix GC DMA commands (#3976)
GC p4 had a typo for the type of DMA command (mem2mem instead of
phv2mem). This results in an invalid address being used in DMA commands.
The problem was caught during RTL test with real hbm, but the bug has
been there a long time.

Also fix bug in GC code that does not skip the scratch area of the
descriptor when reading the page address.