Optimizations for UI compile: (#4064)
* Optimizations for UI compile:

With these optimizations, UI compile  on docker for mac is down to 2
minutes on a fresh tree from 8 minutes presently. On Linux machines its
much faster.

Changes:
In the docker container, create a tar ball and copy the tar ball to the
workspace instead of copying each file individually.
On the workspace, recopy the .tgz file only of the file does not exist
already.

* fix alignment of lines. add echo for the docker commands to be run