[Review-2] Include Netapp review comments
Per external and internal reviews (04/17 & 04/23) on the API header,
this commit:
	- removes 'pad' as a service
	= adds 'zero_pad' to compression desc
	- adds 'insert_header' to compression desc
	- changes the order of enum
	- adds 'pnso_init_params'
	- adds 'interim_buf' to pnso_service_result
	- modifies hash/chksum reporting members in pnso_service_result

Instead of using 'pad' as a service, a boolean flag is being introduced in
'compression descriptor' to fold it within the compmression service.
NOTE: 4KB is assumed to be the default size of a block.

'zero_bad' and 'insert_header' may need to be changed to bit-flags.

Signed-off-by: Mohan Kumar <mohan@pensando.io>