Refactor GC code so it doesn't change on adding producer (#4201)
Currently GC code needs to be modified every time a new producer to GC
is added, since each producer was a unique ring. Refactored the code so
the producer is now a qid and all producer related information is in the
CB. This ensures the GC code does not need to be touched when adding a
new producer, but it will still involve changes to the CB/qstate setup
etc.