Fixes #4113 (#4231)
`TestMockElasticServerRestart` was failing due to excessive
elasticsearch restarts before the client could re-estalish the broken
connection. Test restarts elasticsearch every 500ms and the client tries
to re-establish connection in 1s interval. So, by the time the
connection was reset, there was already a restart in progress which was
causing the caller to exhaust all the retries and that resulted in test
failure. To avoid this issue, this fix increases the restart interval so,
that once the client is reset, there will be some time to get the test
going (elastic operations).

Signed-off-by: Yuva Shankar <yuva29@users.noreply.github.com>