tools: Add checkpatch.pl to follow Linux kernel coding guidelines
checkpatch.pl and other related Linux coding guidelines tools come along
with Linux kernel build environment. However, we would need this tool to
be run against the source code that is being developed for Linux kernel.

This commit adds this tool temporarily until the build environment is
settled.  The script is taken as-is from linux-4.15.9 code base.

Signed-off-by: Mohan Kumar <mohan@pensando.io>