[Review-2.1] Add flag to produce hash/checksum even if CP fails
Per internal discussion on 04/30, we may need to cater a use case.
That is in a compression followed by hash and/or checksum service chain,
even if compression fails, the hash/checksum service should continue to
operate on the source input buffer to produce the hash(es) and/or
checksum(s).

This commit adds a flag in CP descriptor to notify the hash/checksum
service to make use of the input source buffer as the source for
hash/checksum computations.

Signed-off-by: Mohan Kumar <mohan@pensando.io>