pnso_api: [Review-2.5] Beautify WAFL proto to align to checksum/hash/dst split
Existing code was generically using 'tags' member to refer to hash or
checksum. Due to the change in the structure definitions in the external
API, and as the WAFL proto code supports only hash service as of today,
for now this commit replaces 'tags' member with 'hash_tags' to be
specific.

Also this commit sets the destination sgl only to compress and decrypt
service.

Signed-off-by: Mohan Kumar <mohan@pensando.io>