linux ionic: make devcmds interruptible and increase devcmd timeout to 30s
Make devcmd interruptible.  All devcmds are called in process context,
so while sleeping while waiting for devcmd to complete, sleep in such
a way that the user can interrupt (^C) the operation, aborting the devcmd.

Devcmd timeouts weren't consistent in driver, so make them all 30secs.

We could make devcmds timeout infinite, but that would not be acceptable
upstream, so waiting 30secs is a reasonable compromise.  If devcmds are
taking longer than 30secs to complete, we probably have other problems.

Also adds a module parameter to override the default 30secs devcmd timeout
to some other value.  Only available on HAPS version of driver for now.