venice image. rest endpoint to trigger node upgrade. (#4787)
* debug rest endpoint to load the new containers from an image file

Now we have a Naples image
    a tar file with all the contents of various images and json

debug rest handler which loads the images from the new image and runs
the commands specified in json file to run the installation steps.
    till minio is integrated, it just assumes that image file is at
    /var/log/pensando

cmd container is packaged with docker.
    needed to load all the docker images.
    Its also started with docker.sock from the host

Added tryRestart to systemd which restarts a service only if running.

Add back pen-filebeatconfig

the image has INSTALL.sh contained in it. This specifes the steps to be
run on fresh installation.

* implement Restart in systemd mock

* Fix node-cleanup to acount for new layout in /etc/pensando. Dont hardcode ip address in test/e2e/dind/do.py

* convert cluster tests to Eventually. Especially after a node delete tests, it takes few seconds for the services to start running after cleanup. Dont fail in such cases..

* create /var/log/pensando and mount it as part of initscript

* review comments - propogate the error code up

* for the failover tests 2 nodes are enough. dont need 3.

* /host/etc/pensando/configs support.

/host/etc/pensando/configs directory on venice hosts all the config file
which are generated from builds.
These are different from config files that are generated at run-time
based on user-configs. (Such configs need to be cleaned up as part of
node disjoin)

Restore the build-time generated artifacts as part of node-cleanup and
remove all other generated config files.

* List the smartnics in each iteration