Fix intermittent TestCrudOps failure (#4894)
* Fix intermittent TestCrudOps failure

fixes #4712
sometimes the watch event may start after some of hte test crud ops have
been processed. The watch event q in such conditions looks up the
current store and lists all the current objects before starting the
incremental events watch. The list from the store may not be in the same
order as what the test is expecting. This change makes that
deterministic to help with tests. This does not change the contract to
clients

* review comments