Add retries to start/stop tests (#5207)
There are few tests that make sure that a server cleans up properly
when it exits and then it can be restarted. Recently we have seen
failures due to the fact that on a busy system it may take some time
for the OS to free up a TCP port after the test invokes Stop() on a grpc
server. This change makes the test perform multiple attempts to restart
the server before reporting an error.