osal_logger: Retire logger test for now
Earlier osal_logger was exercising user space log macros under the
'test' folder.  As the macros and logger routines are being exercised
in osal_test, and verified via sanity tests, this commit parks aside
this test utility. For log level verification, we can rely on 'dmseg'
log level flag.  And, if/when new features are added in the future,
this test may be used to add tests/UTs.

Signed-off-by: Mohan Kumar <mohan@pensando.io>