pnso_cp_ut: Add UTs
This commit includes basic set of UTs for the service ops (i.e.
setup/teardown/etc.) -- given the current phase of the code:
	- not being tested against model or real hardware
	- handling only single request
	- single service
	- sync mode
some tweaks are made to exercise as many code paths as possible.
Additional UTs will be added when the respective layers are added,
and when more code changes get in.

Signed-off-by: Mohan Kumar <mohan@pensando.io>