Fix a race condition in CMD (#5474)
Fixes https://github.com/pensando/sw/issues/5434

The race condition happens when a CMD instance receives a cluster
disjoin request while it is still processing a cluster create/join request.

The Disjoin handler closes the CertMgr instance and sets the env reference to nil,
but there might be pending deferred operations that need to be handled gracefully.

This race condition was triggered by the fact that on a busy system etcd can take
time to bind to the TCP port and accepting client connections after it has been
started. Previous timeout was at 5 seconds, this PR bumps it up to 15.