Make swagger-gen stable wrt any.proto (#5561)
* Vendor files

* Make swagger-gen stable wrt any.proto

proto compiler resolves any.proto to local /usr/include/.. version when
none of the -I paths are able to provide a proto import. This is an
issue with native protos provided by the protoc installation like
any.proto

The PR changes this to import from a known location in the codebase.

for this we need to prevet dep from the pruning out the protos dir  supplied
by gogoproto. But preventing dep from  pruning specific paths when the path
does not have a go package seems to not work. Hence preventing pruning at a
more coarse level.