pnso_cpdc: Get per_core_resource via lif to access sgl/chain/entry mpools
This commit exports a function to have pnso api access to lif, so that
it can obtain the per core resource, and then access to
sgl/chain/chain_entry mpools.  While accessing the global variable (i.e.
lif) is one thing, the callers of these sub-routines need to be
refactored also to better structure the routines.

Signed-off-by: Mohan Kumar <mohan@pensando.io>