Locally issued certificates for NAPLES processes (#5713)
* Locally issued certificates for NAPLES processes

This change introduces a proxy functionality that allows NAPLES
processes to request their mTLS certificates to a local endpoint
hosted by NMD instead of Venice CMD endpoint. This is necessary
because Venice CMD does not have any mean to validate a certificate
request coming from a remote process.

When NMD starts, it generates a private key and a CSR and submits
it in the registration request sent to the CMD unauthenticated port.
If the NIC is admitted, CMD signs the CSR and returns the certificate
in the registration response. Once NMD has received the certificate,
it uses it to open a mTLS connection to the CMD authenticated port.
At this point NMD also starts a certificate proxy that receives
certificates requests from local processes (same as on Venice)
and forwards them to CMD for signing over the secure channel.

Other changes:
- Includes minor cleanups in the Certificates API to improve consistency.
- New keymgr_based TLS provider to store the cluster key and cert in NMD and
  use them to establish the secure channel to CMD
- Enhancements in KeyMgr to support "update" functions (preparatory work
  to support key rotation)
- Removed naples_tls_provider.go as it is no longer needed

* Autogenerated files for https://github.com/pensando/sw/pull/5713